add story with my location

+++ app/src/main/java/me/varoa/sad/core/domain/model/NewStory.kt
@@ -5,6 +5,6 @@ import java.io.File
 data class NewStory(
     val description: String,
     val photo: File,
-    val lat: Float? = null,
-    val lon: Float? = null
+    val lat: Double? = null,
+    val lon: Double? = null
 )
